Moving fast api to approapriate services
