+++ app/api/streamsocket.py
@@ -1,187 +1,15 @@
 import asyncio
-
-import aioredis
 import uuid
-from aioredis.errors import \
-    ConnectionClosedError as ServerConnectionClosedError
+from starlette.websockets import WebSocket
+
 from fastapi import Depends
 from fastapi.routing import APIRouter
 from loguru import logger
-from starlette.websockets import WebSocket, WebSocketDisconnect
-from websockets.exceptions import ConnectionClosedError, ConnectionClosedOK
 
-from app.core.config import (NUM_PREVIOUS, REDIS_HOST, REDIS_PORT,
-                             STREAM_MAX_LEN, XREAD_COUNT, XREAD_TIMEOUT, ALLOWED_ROOMS)
-import contextvars
+from app.services.authentication import verify_user_for_room
+from app.services.stream_rw import ws_recieve, ws_send
 
 router = APIRouter()
-
-
-async def verify_user_for_room(chat_info):
-    verified = True
-    pool = await get_redis_pool()
-    if not pool:
-        logger.info('Redis connection failure')
-        return False
-    # check for duplicated user names
-    already_exists = await pool.sismember(chat_info['users'], chat_info['username'])
-    if already_exists:
-        logger.error(chat_info['username'] +' user already_exists in ' + chat_info['room'])
-        verified = False
-    # TODO: add more validation here
-    # whitelist rooms
-    if not chat_info['room'] in ALLOWED_ROOMS:
-        verified = False
-    pool.close()
-    return verified
-
-
-async def get_redis_pool():
-    try:
-        pool = await aioredis.create_redis_pool(
-            (REDIS_HOST, REDIS_PORT), encoding='utf-8')
-        return pool
-    except ConnectionRefusedError as e:
-        logger.info('cannot connect to redis on:', REDIS_HOST, REDIS_PORT)
-        return None
-
-
-async def ws_send(websocket: WebSocket, chat_info: dict):
-    '''
-    wait for new items on chat stream and
-    send data from server to client over a WebSocket
-
-    :param websocket:
-    :type websocket:
-    :param chat_info:
-    :type chat_info:
-    '''
-    pool = await get_redis_pool()
-    latest_ids = ['$']
-    ws_connected = True
-    first_run = True
-    while pool and ws_connected:
-        try:
-            if first_run:
-                # fetch some previous chat history
-                events = await pool.xrevrange(
-                    stream=chat_info['stream'],
-                    count=NUM_PREVIOUS,
-                    start='+',
-                    stop='-'
-                )
-                first_run = False
-                events.reverse()
-                for e_id, e in events:
-                    e['e_id'] = e_id
-                    await websocket.send_json(e)
-            else:
-                events = await pool.xread(
-                    streams=[chat_info['stream']],
-                    count=XREAD_COUNT,
-                    timeout=XREAD_TIMEOUT,
-                    latest_ids=latest_ids
-                )
-                for _, e_id, e in events:
-                    e['e_id'] = e_id
-                    await websocket.send_json(e)
-                    latest_ids = [e_id]
-        except ConnectionClosedError:
-            ws_connected = False
-
-        except ConnectionClosedOK:
-            ws_connected = False
-
-        except ServerConnectionClosedError:
-            logger.info('redis server connection closed')
-            return
-    pool.close()
-
-
-async def ws_recieve(websocket: WebSocket, chat_info: dict):
-    '''
-    receive json data from client over a WebSocket, add messages onto the
-    associated chat stream
-
-    :param websocket:
-    :type websocket:
-    :param chat_info:
-    :type chat_info:
-    '''
-
-    ws_connected = False
-    pool = await get_redis_pool()
-    added = await add_room_user(chat_info, pool)
-
-    if added:
-        await announce(pool, chat_info, 'connected')
-        ws_connected = True
-    else:
-        logger.info('duplicate user error')
-
-    while ws_connected:
-        try:
-            data = await websocket.receive_json()
-            #logger.info(data)
-            if type(data) == list and len(data):
-                data = data[0]
-            fields = {
-                'uname': chat_info['username'],
-                'msg': data['msg'],
-                'type': 'comment',
-                'room': chat_info['room'],
-            }
-            await pool.xadd(stream=chat_info['stream'],
-                            fields=fields,
-                            message_id=b'*',
-                            max_len=STREAM_MAX_LEN)
-        except WebSocketDisconnect:
-            await remove_room_user(chat_info, pool)
-            await announce(pool, chat_info, 'disconnected')
-            ws_connected = False
-
-        except ServerConnectionClosedError:
-            logger.info('redis server connection closed')
-            return
-
-        except ConnectionRefusedError:
-            logger.info('redis server connection closed')
-            return
-
-    pool.close()
-
-
-async def add_room_user(chat_info: dict, pool):
-    return await pool.sadd(chat_info['users'], chat_info['username'])
-
-
-async def remove_room_user(chat_info: dict, pool):
-    return await pool.srem(chat_info['users'], chat_info['username'])
-
-
-async def room_users(chat_info: dict, pool):
-    users = await pool.smembers(chat_info['users'])
-    logger.info(len(users))
-    return users
-
-
-async def announce(pool, chat_info: dict, action: str):
-    '''
-    add an announcement event onto the redis chat stream
-    '''
-    users = await room_users(chat_info, pool)
-    fields = {
-        'msg': f"{chat_info['username']} {action}",
-        'action': action,
-        'type': 'announcement',
-        'users': ', '.join(users),
-        'room': chat_info['room'],
-    }
-
-    await pool.xadd(stream=chat_info['stream'],
-                    fields=fields,
-                    message_id=b'*',
-                    max_len=STREAM_MAX_LEN)
 
 
 async def chat_info_vars(username: str = None, room: str = None):
