+++ app/db/connection.py
@@ -8,8 +8,5 @@ logger = logging.getLogger(__name__)
 
 
 async def get_redis_pool():
-    pool = await aioredis.create_redis_pool(
+    return await aioredis.create_redis_pool(
         (REDIS_HOST, REDIS_PORT), encoding='utf-8')
-    yield pool
-    pool.close()
-    await pool.wait_closed()
